Update entities and make them workgit add .

+++ app/src/main/java/fr/cyberix/kolo/activities/LoginActivity.java
@@ -18,6 +18,7 @@ import android.widget.ProgressBar;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import java.math.BigDecimal;
 import java.util.Calendar;
 import java.util.Date;
 
@@ -114,9 +115,9 @@ public class LoginActivity extends AppCompatActivity {
 			loginAttempt.setSubscriberId(telInfo.getSubscriberId());
 			loginAttempt.setDeviceId(telInfo.getDeviceId());
 			if (loc != null) {
-				loginAttempt.setLatitude(Double.valueOf(loc.getLatitude()));
-				loginAttempt.setLongitude(Double.valueOf(loc.getLongitude()));
-				loginAttempt.setAccuracy(Double.valueOf(loc.getAccuracy()));
+				loginAttempt.setLatitude(BigDecimal.valueOf(loc.getLatitude()));
+				loginAttempt.setLongitude(BigDecimal.valueOf(loc.getLongitude()));
+				loginAttempt.setAccuracy(BigDecimal.valueOf(loc.getAccuracy()));
 			}
 		}
 		
@@ -217,9 +218,9 @@ public class LoginActivity extends AppCompatActivity {
 		loginAttempt.setSubscriberId(telInfo.getSubscriberId());
 		loginAttempt.setDeviceId(telInfo.getDeviceId());
 		if (loc != null) {
-			loginAttempt.setLatitude(Double.valueOf(loc.getLatitude()));
-			loginAttempt.setLongitude(Double.valueOf(loc.getLongitude()));
-			loginAttempt.setAccuracy(Double.valueOf(loc.getAccuracy()));
+			loginAttempt.setLatitude(BigDecimal.valueOf(loc.getLatitude()));
+			loginAttempt.setLongitude(BigDecimal.valueOf(loc.getLongitude()));
+			loginAttempt.setAccuracy(BigDecimal.valueOf(loc.getAccuracy()));
 		}
 		ServiceHelper.doInBackground(loginAttempt, tryToLogIn);
 //		userSignInTask = new UserSignInTask();
